requirements.txt
amazon-textract-textractor==1.8.2
amazon-textract-response-parser==1.0.3
amazon-textract-caller==0.2.4
boto3==1.34.148
botocore==1.34.148
nltk==3.8.1
pandas==2.2.2